Research of bank borrowers
The customer is the credit department of the bank. It is necessary to find out whether the marital status and the number of children of the client affect the fact of repaying the loan on time. Input data from the bank - statistics on the solvency of customers. The results of the study will be taken into account when building a credit scoring model - a special system that assesses the ability of a potential borrower to return a loan to a bank.

The dataframe contains:

children — number of children in the family days_employed - total work experience in days dob_years — client's age in years education — client education level education_id — education level ID family_status — marital status family_status_id — marital status identifier gender - gender of the client income_type — employment type debt - whether he had a debt on the return of credits total_income — monthly income purpose — the purpose of obtaining a loan